Add files via upload
Speech Emotion Recognition (SER) is one of the most challenging tasks in the speech signal analysis domain, it is a research area problem that tries to infer the emotion from the speech signals. Here is the RAVDESS dataset that has been used as the input source (for Audio Datasets). Librosa library has been used for feature extraction.